W142 audit: restore code2 FUN_0077F710 marker (documented boundary-data floor, not a match)

+++ src/Event/Fcl/fcl_misc.c
@@ -2007,7 +2007,7 @@ void FUN_003ca960(int param_1)
 
   int iVar4;
 
-  long lVar5;
+  int lVar5;
 
   int iVar6;
 
@@ -2053,7 +2053,7 @@ void FUN_003ca960(int param_1)
 
   if ((*(u32 *)(param_1 + 8) & 0x40) != 0) {
 
-    lVar5 = (long)*(short *)(param_1 + 0x16);
+    lVar5 = (int)*(short *)(param_1 + 0x16);
 
     fVar7 = (float)(int)*(short *)(param_1 + 0x14);
 
@@ -2063,7 +2063,7 @@ void FUN_003ca960(int param_1)
 
     if (*(char *)(param_1 + 0xe8) == '\0') {
 
-      lVar5 = (long)((int)*(short *)(param_1 + 0x14) - (int)*(short *)(param_1 + 0x16));
+      lVar5 = (int)*(short *)(param_1 + 0x14) - (int)*(short *)(param_1 + 0x16);
 
       iVar6 = param_1 + 0x70;
